+++ src/main/java/com/example/bank/security/auth/AuthenticationService.java
@@ -1,5 +1,6 @@
 package com.example.bank.security.auth;
 
+import com.example.bank.client.ChangePasswordRequest;
 import com.example.bank.client.ClientRequest;
 import com.example.bank.client.jpa.Client;
 import com.example.bank.client.jpa.ClientRepository;
@@ -23,6 +24,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
+import java.security.Principal;
 import java.time.LocalDate;
 import java.time.Period;
 import java.util.List;
@@ -118,5 +120,18 @@ public class AuthenticationService {
                 new ObjectMapper().writeValue(response.getOutputStream(), authResponse);
             }
         }
+    }
+
+    public void changePassword(ChangePasswordRequest request, Principal connectedUser) {
+        Client client = (Client) ((UsernamePasswordAuthenticationToken) connectedUser).getPrincipal();
+        if (!passwordEncoder.matches(request.currentPassword(), client.getPassword())) {
+            throw new IllegalStateException("Wrong password.");
+        }
+        if (!request.newPassword().equals(request.confirmationPassword())) {
+            throw new IllegalStateException("Passwords are not the same.");
+        }
+        client.setPassword(passwordEncoder.encode(request.newPassword()));
+        clientRepository.save(client);
+        log.info("Password has been successfully changed.");
     }
 }
