+++ src/test/java/com/example/bank/integration/DemoAppTests.java
@@ -1,19 +1,16 @@
 package com.example.bank.integration;
 
-import com.example.bank.BankApplication;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
+import org.springframework.context.annotation.Profile;
+import org.springframework.test.context.jdbc.Sql;
 
-@SpringBootTest(classes = BankApplication.class,
-        webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@ActiveProfiles("test")
+@SpringBootTest
 public class DemoAppTests {
     @Test
     void contextLoads() {
     }
 
-    //    @Sql({"classpath:schema.sql"/*, "classpath:data.sql"*/})
     @Test
     void something() {
         System.out.println("test test test test");
